requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
Django==4.1.5
django-extensions==3.2.1
django-sqlformatter==1.0.3
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==2.0.0
mysqlclient==2.1.1
packaging==23.0
pluggy==1.0.0
Pygments==2.14.0
pytest==7.2.1
pytest-django==4.5.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
